requirements.txt
appdirs==1.4.3
astroid==2.4.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
parse==1.15.0
Pillow==7.1.2
pyee==7.0.1
pylint==2.5.2
pyppeteer==0.2.2
pyquery==1.4.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
toml==0.10.0
tqdm==4.46.0
typed-ast==1.4.1
urllib3==1.25.9
w3lib==1.21.0
websockets==8.1
wrapt==1.12.1
